+++ src/main/java/chess/repository/ChessRepository.java
@@ -1,32 +1,61 @@
 package chess.repository;
 
+import chess.dao.PieceDao;
+import chess.dao.TurnDao;
 import chess.dto.request.ChessRequestDto;
 import chess.dto.request.MoveRequestDto;
 import chess.dto.request.TurnChangeRequestDto;
 import chess.dto.request.TurnRequestDto;
 import org.springframework.stereotype.Repository;
 
-import java.sql.SQLException;
 import java.util.List;
 import java.util.Map;
 
 @Repository
-public interface ChessRepository {
-    void initializePieceStatus(final Map<String, String> board) throws SQLException;
+public class ChessRepository {
+    private final PieceDao pieceDao;
+    private final TurnDao turnDao;
 
-    void initializeTurn() throws SQLException;
+    public ChessRepository(PieceDao pieceDao, TurnDao turnDao) {
+        this.pieceDao = pieceDao;
+        this.turnDao = turnDao;
+    }
 
-    List<ChessRequestDto> showAllPieces() throws SQLException;
+    public void initializePieceStatus(final Map<String, String> board) {
+        for (Map.Entry<String, String> boardStatus : board.entrySet()) {
+            pieceDao.initializePieceStatus(boardStatus.getValue(), boardStatus.getKey());
+        }
+    }
 
-    List<TurnRequestDto> showCurrentTurn() throws SQLException;
+    public void initializeTurn() {
+        turnDao.initializeTurn();
+    }
 
-    void movePiece(final MoveRequestDto moveRequestDto) throws SQLException;
+    public List<ChessRequestDto> showAllPieces() {
+        return pieceDao.showAllPieces();
+    }
 
-    void changeTurn(final TurnChangeRequestDto turnChangeRequestDto) throws SQLException;
+    public List<TurnRequestDto> showCurrentTurn() {
+        return turnDao.showCurrentTurn();
+    }
 
-    void removeAllPieces() throws SQLException;
+    public void movePiece(final MoveRequestDto moveRequestDto) {
+        pieceDao.movePiece(moveRequestDto);
+    }
 
-    void removeTurn() throws SQLException;
+    public void changeTurn(final TurnChangeRequestDto turnChangeRequestDto) {
+        turnDao.changeTurn(turnChangeRequestDto);
+    }
 
-    void removePiece(final MoveRequestDto moveRequestDto) throws SQLException;
+    public void removeAllPieces() {
+        pieceDao.removeAllPieces();
+    }
+
+    public void removeTurn() {
+        turnDao.removeTurn();
+    }
+
+    public void removePiece(final MoveRequestDto moveRequestDto) {
+        pieceDao.removePiece(moveRequestDto);
+    }
 }
